Reorganize text expansion callout into smaller component files
